+++ changelog.md
@@ -4,6 +4,8 @@
 * Added Creeper Collateral option. Disabled by default. If enabled will make creepers' explosion drop no blocks.
 * Updated effects icons
 * Food is now eaten slightly faster
+* Freeze no longer applies injured
+* Effects are no longer applied if duration is 0
 
 ## 2.11.4
 * Requires InsaneLib 1.5.1. Fix tags not working for blacklists.